chore: 컴포넌트 수정 및 파일 삭제
